Significant improvement in the function of the structure making plugin for hyde
The plugin is now called 'presenter' and it is very generalized so as to be useful in many situations when wanting
to build a series of html structures that contain text. These lists of structures are called 'presentations' and the
individual elements are called 'slides'. Site config, resource.meta, and single element configurations are possible.
id and class enumeration is also possible. This commit does not contain a good example, but 2 very basic simple examples